requirements.txt
Django==1.10.5
djangorestframework==3.5.3
MySQL-python==1.2.5
Pillow
beautifulsoup4==4.5.3
django-cron==0.5.0
selenium==3.0.2
python-dateutil

# apt-get install nodejs
# apt-get install nodejs-legacy
# npm -g install phantomjs-prebuilt

# 20 10 * * * source /home/ubuntu/.bashrc && source /home/deepak/Desktop/django_alireza/alireza/bin/activate && python /home/deepak/Desktop/django_alireza/alireza/src/manage.py runcrons > /home/ubuntu/cronjob.log